requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.4.0
celery==5.4.0
redis==5.0.7
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9
